Rename file add custom exception for yaml config error
