+++ brats/exceptions.py
@@ -1,5 +1,15 @@
+from dacite import DaciteError
+
+
 class CPUNotCompatibleException(Exception):
     """Exception raised when an CPU-incompatible algorithm tries to run on CPU."""
+
+    def __init__(self, message: str):
+        super().__init__(message)
+
+
+class AlgorithmConfigException(DaciteError):
+    """Exception raised when the algorithm config file has issues."""
 
     def __init__(self, message: str):
         super().__init__(message)
